Modified schedPeriod type to reduce memory occupation

+++ project/arduino/smart_waste_disposal/include/communication_task.h
@@ -5,7 +5,7 @@
 
 class CommunicationTask: public Task {
     public:
-        void step(long schedPeriod);
+        void step(int schedPeriod);
 };
 
 #endif
