+++ project/arduino/smart_waste_disposal/include/container_handler_task.h
@@ -5,7 +5,7 @@
 
 class ContainerHandlerTask: public Task {
     public:
-        void step(long schedPeriod);
+        void step(int schedPeriod);
 };
 
 #endif
