+++ project/arduino/smart_waste_disposal/include/temp_monitor_task.h
@@ -5,7 +5,7 @@
 
 class TempMonitorTask: public Task {
     public:
-        void step(long schedPeriod);
+        void step(int schedPeriod);
 };
 
 #endif
